% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{CalculateRespiratoryMetrics}
\alias{CalculateRespiratoryMetrics}
\title{Metric calculator}
\usage{
CalculateRespiratoryMetrics(data, file_path, observer)
}
\arguments{
\item{data}{A data frame containing respiratory data.}

\item{file_path}{The file path of the data source.}

\item{observer}{The name of the observer.}
}
\value{
A data frame with calculated respiratory metrics.
}
\description{
this function calculate breathing metrics
}
\examples{
# example code
file_path <- "https://raw.githubusercontent.com/noromalala/data_tsv/master/Sarah.txt"
observer <- "Sarah"
respiratory_data <- read.table(file_path, header = TRUE)
metrics <- CalculateRespiratoryMetrics(respiratory_data, file_path, observer)
print(metrics)
}
